renaming files + small css changes
