+++ Frontend/src/App.js
@@ -1,9 +1,5 @@
 import React, { useState, useRef, useEffect } from 'react';
-import LandingPageComponent from './components/LandingPageComponent';
-import FormComponent from './components/FormComponent';
-import CalendarComponent from './components/CalendarComponent';
-import ErrorPage from './components/ErrorPage';
-import Footer from './components/Footer'
+import { Calendar, ErrorPage, Footer, Form, LandingPage } from './components'
 
 const App = () => {
   const [displayCalendar, setDisplayCalendar] = useState(false);
@@ -19,7 +15,7 @@ const App = () => {
     }
   });
 
-  const displayCalendarComponent = () => (displayCalendar ? <CalendarComponent title={term} events={schedules} scheduleCount={scheduleCount} setScheduleCount={setScheduleCount} ref={calendarRef} /> : null);
+  const displayCalendarComponent = () => (displayCalendar ? <Calendar title={term} events={schedules} scheduleCount={scheduleCount} setScheduleCount={setScheduleCount} ref={calendarRef} /> : null);
 
   if (serverError) {
     return <ErrorPage />
@@ -27,8 +23,8 @@ const App = () => {
 
   return (
     <div>
-      <LandingPageComponent />
-      <FormComponent setDisplayCalendar={setDisplayCalendar} setTerm={setTerm} setSchedules={setSchedules} setScheduleCount={setScheduleCount} setServerError={setServerError} />
+      <LandingPage />
+      <Form setDisplayCalendar={setDisplayCalendar} setTerm={setTerm} setSchedules={setSchedules} setScheduleCount={setScheduleCount} setServerError={setServerError} />
       {displayCalendarComponent()}
       <Footer />
     </div>
