+++ Frontend/src/common/utils.js
@@ -1,5 +1,3 @@
-// Formatting functions 
-
 export const parseInputs = (inputs) => {
     const courses = [
         inputs.course1,
