+++ Frontend/src/components/Calendar.js
@@ -2,14 +2,11 @@ import React, { useState } from 'react';
 import FullCalendar from '@fullcalendar/react';
 import timeGridPlugin from '@fullcalendar/timegrid';
 import rrulePlugin from '@fullcalendar/rrule';
-import '../styles/CalendarComponent.css';
 import { getCourseTime } from '../common/utils';
-import Snackbar from '@mui/material/Snackbar';
-import MuiAlert from '@mui/material/Alert';
-import Box from '@mui/material/Box';
-import Modal from '@mui/material/Modal';
+import { Box, Modal, MuiAlert, Snackbar } from '@mui/material';
+import '../styles/Calendar.css';
 
-const MyCalendar = React.forwardRef(({ title, events, scheduleCount, setScheduleCount }, ref) => {
+const Calendar = React.forwardRef(({ title, events, scheduleCount, setScheduleCount }, ref) => {
     const [openAlert, setOpenAlert] = useState(false);
     const [openModal, setOpenModal] = useState(false);
     const [alertMessage, setAlertMessage] = useState('');
@@ -111,7 +108,7 @@ const MyCalendar = React.forwardRef(({ title, events, scheduleCount, setSchedule
                 allDaySlot={false}
                 slotMinTime={"8:00:00"}
                 slotMaxTime={"21:00:00"}
-                dayHeaderFormat={{ weekday: 'long' }}
+                dayHeaderFormat={{ weekday: 'short' }}
                 height="auto"
                 initialDate={earliestStartDate(events)}
                 events={events[scheduleCount]["sync"]}
@@ -151,4 +148,4 @@ const MyCalendar = React.forwardRef(({ title, events, scheduleCount, setSchedule
     );
 });
 
-export default MyCalendar;
+export default Calendar;
