+++ Frontend/src/styles/Form.css
@@ -6,7 +6,7 @@
     padding: 20px;
     border: thin solid black;
     border-radius: 9px;
-    margin-bottom: 350px;
+    margin-bottom: 150px;
     color: black;
 }
